requirements.txt
beautifulsoup4==4.11.1
easysnmp==0.2.6
mysql-connector-python==8.3.0
pysftp==0.2.9
